+++ cx/ast/garbage_collector.go
@@ -1,7 +1,7 @@
-package cxcore
+package ast
 
 import (
-	"github.com/skycoin/cx/cx/ast"
+	"github.com/skycoin/cx/cx"
 	"github.com/skycoin/cx/cx/constants"
 	"github.com/skycoin/cx/cx/helper"
 	"github.com/skycoin/cx/cx/mem"
@@ -10,7 +10,7 @@ import (
 )
 
 // MarkAndCompact ...
-func MarkAndCompact(prgrm *ast.CXProgram) {
+func MarkAndCompact(prgrm *CXProgram) {
 	var fp int
 	var faddr = int32(constants.NULL_HEAP_ADDRESS_OFFSET)
 
@@ -135,7 +135,7 @@ func MarkAndCompact(prgrm *ast.CXProgram) {
 }
 
 // updateDisplaceReference performs the actual addition or subtraction of `plusOff` to the address being pointed by the element at `atOffset`.
-func updateDisplaceReference(prgrm *ast.CXProgram, updated *map[int]int, atOffset, plusOff int) {
+func updateDisplaceReference(prgrm *CXProgram, updated *map[int]int, atOffset, plusOff int) {
 	// Checking if it was already updated.
 	if _, found := (*updated)[atOffset]; found {
 		return
@@ -155,7 +155,7 @@ func updateDisplaceReference(prgrm *ast.CXProgram, updated *map[int]int, atOffse
 }
 
 // doDisplaceReferences checks if the element at `atOffset` is pointing to an object on the heap and, if this is the case, it displaces it by `plusOff`. `updated` keeps a record of all the offsets that have already been updated.
-func doDisplaceReferences(prgrm *ast.CXProgram, updated *map[int]int, atOffset int, plusOff int, baseType int, declSpecs []int) {
+func doDisplaceReferences(prgrm *CXProgram, updated *map[int]int, atOffset int, plusOff int, baseType int, declSpecs []int) {
 	var numDeclSpecs = len(declSpecs)
 
 	// Getting the offset to the object in the heap.
@@ -190,7 +190,7 @@ func doDisplaceReferences(prgrm *ast.CXProgram, updated *map[int]int, atOffset i
 			(numDeclSpecs == 1 && baseType == constants.TYPE_STR) {
 			// Then we need to iterate each of the slice objects
 			// and check if we need to update their address.
-			sliceLen := helper.Deserialize_i32(GetSliceHeader(heapOffset + int32(condPlusOff))[4:8])
+			sliceLen := helper.Deserialize_i32(cxcore.GetSliceHeader(heapOffset + int32(condPlusOff))[4:8])
 
 			offsetToElements := constants.OBJECT_HEADER_SIZE + constants.SLICE_HEADER_SIZE
 
@@ -210,7 +210,7 @@ func doDisplaceReferences(prgrm *ast.CXProgram, updated *map[int]int, atOffset i
 }
 
 // DisplaceReferences displaces all the pointer-like variables, slice elements or field structures by `off`. `numPkgs` tells us the number of packages to consider for the reference desplacement (this number should equal to the number of packages that represent the blockchain code in a CX chain).
-func DisplaceReferences(prgrm *ast.CXProgram, off int, numPkgs int) {
+func DisplaceReferences(prgrm *CXProgram, off int, numPkgs int) {
 	// We're going to keep a record of all the references that were already updated.
 	updated := make(map[int]int)
 
@@ -238,7 +238,7 @@ func DisplaceReferences(prgrm *ast.CXProgram, off int, numPkgs int) {
 }
 
 // Mark marks the object located at `heapOffset` as alive and sets the object's referencing address to `heapOffset`.
-func Mark(prgrm *ast.CXProgram, heapOffset int32) {
+func Mark(prgrm *CXProgram, heapOffset int32) {
 	// Marking as alive.
 	prgrm.Memory[heapOffset] = 1
 
@@ -247,7 +247,7 @@ func Mark(prgrm *ast.CXProgram, heapOffset int32) {
 }
 
 // MarkObjectsTree traverses and marks a possible tree of heap objects (slices of slices, slices of pointers, etc.).
-func MarkObjectsTree(prgrm *ast.CXProgram, offset int, baseType int, declSpecs []int) {
+func MarkObjectsTree(prgrm *CXProgram, offset int, baseType int, declSpecs []int) {
 	lenMem := len(prgrm.Memory)
 	// Checking if it's a valid heap address. An invalid address
 	// usually occurs in CX chains, with the split of blockchain
@@ -281,7 +281,7 @@ func MarkObjectsTree(prgrm *ast.CXProgram, offset int, baseType int, declSpecs [
 				declSpecs[1] == constants.DECL_POINTER)) ||
 			(numDeclSpecs == 1 && baseType == constants.TYPE_STR) {
 			// Then we need to iterate each of the slice objects and mark them as alive
-			sliceLen := helper.Deserialize_i32(GetSliceHeader(heapOffset)[4:8])
+			sliceLen := helper.Deserialize_i32(cxcore.GetSliceHeader(heapOffset)[4:8])
 
 			for c := int32(0); c < sliceLen; c++ {
 				offsetToElements := constants.OBJECT_HEADER_SIZE + constants.SLICE_HEADER_SIZE
@@ -299,13 +299,13 @@ func MarkObjectsTree(prgrm *ast.CXProgram, offset int, baseType int, declSpecs [
 }
 
 // updatePointer changes the address of the pointer located at `atOffset` to `newAddress`.
-func updatePointer(prgrm *ast.CXProgram, atOffset int, toAddress int32) {
+func updatePointer(prgrm *CXProgram, atOffset int, toAddress int32) {
 	mem.WriteMemI32(prgrm.Memory, atOffset, toAddress)
 }
 
 // updatePointerTree changes the address of the pointer located at `atOffset` to `newAddress` and checks if it is the
 // root of a tree of objects, such as a slice or the instance of a struct where some of its fields are pointers.
-func updatePointerTree(prgrm *ast.CXProgram, atOffset int, oldAddr, newAddr int32, baseType int, declSpecs []int) {
+func updatePointerTree(prgrm *CXProgram, atOffset int, oldAddr, newAddr int32, baseType int, declSpecs []int) {
 	var numDeclSpecs = len(declSpecs)
 
 	// Getting the offset to the object in the heap
@@ -330,7 +330,7 @@ func updatePointerTree(prgrm *ast.CXProgram, atOffset int, oldAddr, newAddr int3
 			(numDeclSpecs == 1 && baseType == constants.TYPE_STR) {
 			// Then we need to iterate each of the slice objects
 			// and check if we need to update their address.
-			sliceLen := helper.Deserialize_i32(GetSliceHeader(heapOffset)[4:8])
+			sliceLen := helper.Deserialize_i32(cxcore.GetSliceHeader(heapOffset)[4:8])
 
 			offsetToElements := constants.OBJECT_HEADER_SIZE + constants.SLICE_HEADER_SIZE
 
@@ -356,7 +356,7 @@ func updatePointerTree(prgrm *ast.CXProgram, atOffset int, oldAddr, newAddr int3
 // For example, if `foo` was pointing to an object located at address 5151 and after calling the garbage collector it was
 // moved to address 4141, every symbol in a `CXProgram`'s `CallStack` and in its global variables need to be updated to point now to
 // 4141 instead of 5151.
-func updatePointers(prgrm *ast.CXProgram, oldAddr, newAddr int32) {
+func updatePointers(prgrm *CXProgram, oldAddr, newAddr int32) {
 	if oldAddr == newAddr {
 		return
 	}
